refactor[util]: refactor util filenames
